Used GetOwner() to grab the object that the script is on
We used GetOwner() to point and get the name of the object that the script is on. We also stored that in an FString,which is a string variable for Unreal, and then called it using the %s in the actual string then referencing the FString OwnerName right after in order to grab the name of the object.